+++ Gravitas.Model/DomainModel/Node/DAO/NodeConstraint.cs
@@ -1,9 +1,10 @@
 using System.ComponentModel.DataAnnotations.Schema;
+using Gravitas.Model.DomainModel.Base;
 
 namespace Gravitas.Model {
 
 	[Table("dev.NodeConstraint")]
-	public partial class NodeConstraint : BaseEntity<long> {
+	public partial class NodeConstraint : BaseEntity<int> {
 		
 		public long NodeId { get; set; }
 		public long? NomenclatureId { get; set; }
