+++ Gravitas.Model/DomainModel/Node/TDO/List/NodeItem.cs
@@ -1,6 +1,8 @@
+using Gravitas.Model.DomainModel.Base;
+
 namespace Gravitas.Model.Dto {
 
-	public class NodeItem : BaseEntity<long> {
+	public class NodeItem : BaseEntity<int> {
 
 		public string Code { get; set; }
 		public string Name { get; set; }
