+++ Gravitas.Model/DomainModel/OpDataState/DTO/Detail/OpDataStateDetail.cs
@@ -1,6 +1,8 @@
+using Gravitas.Model.DomainModel.Base;
+
 namespace Gravitas.Model.Dto {
 	
-	public class OpDataStateDetail : BaseEntity<long> {
+	public class OpDataStateDetail : BaseEntity<int> {
 		
 		public string Name { get; set; }
 	}
