+++ Gravitas.Model/DomainModel/OpDataState/DTO/List/OpDataStateItem.cs
@@ -1,6 +1,8 @@
+using Gravitas.Model.DomainModel.Base;
+
 namespace Gravitas.Model.Dto {
 	
-	public class OpDataStateItem : BaseEntity<long> {
+	public class OpDataStateItem : BaseEntity<int> {
 		
 		public string Name { get; set; }
 	}
