+++ Gravitas.Model/DomainModel/OpRoutine/DAO/OpRoutineProcessor.cs
@@ -1,10 +1,11 @@
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
+using Gravitas.Model.DomainModel.Base;
 
 namespace Gravitas.Model {
 	
-	public partial class OpRoutineProcessor : BaseEntity<long> {
+	public partial class OpRoutineProcessor : BaseEntity<int> {
 
 		public OpRoutineProcessor() {
 			OpRoutineTransitionSet = new HashSet<OpRoutineTransition>();
