+++ Gravitas.Model/DomainModel/OpVisa/DTO/List/OpVisaItem.cs
@@ -1,8 +1,9 @@
 using System;
+using Gravitas.Model.DomainModel.Base;
 
 namespace Gravitas.Model.Dto {
 	
-	public class OpVisaItem : BaseEntity<long> {
+	public class OpVisaItem : BaseEntity<int> {
 
 		public DateTime? DateTime { get; set; }
 		public string Message { get; set; }
