+++ Gravitas.Model/DomainModel/OpVisa/DTO/List/OpVisaItems.cs
@@ -1,8 +1,9 @@
 using System.Collections.Generic;
+using Gravitas.Model.DomainModel.Base;
 
 namespace Gravitas.Model.Dto {
 	
-	public class OpVisaItems : BaseEntity<long> {
+	public class OpVisaItems : BaseEntity<int> {
 
 		public List<OpVisaItem> Items { get; set; }
 	}
