+++ Gravitas.Model/DomainModel/OrganizationUnit/DAO/OrganizationUnitType.cs
@@ -1,11 +1,12 @@
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
+using Gravitas.Model.DomainModel.Base;
 
 namespace Gravitas.Model {
 
 	[Table("OrganizationUnitType")]
-	public partial class OrganizationUnitType : BaseEntity<long> {
+	public partial class OrganizationUnitType : BaseEntity<int> {
 
 		public OrganizationUnitType() {
 			OrganizationUnitSet = new HashSet<OrganizationUnit>();
