+++ Gravitas.Model/DomainModel/OrganizationUnit/DTO/List/OrganizationUnitItem.cs
@@ -1,6 +1,8 @@
+using Gravitas.Model.DomainModel.Base;
+
 namespace Gravitas.Model.Dto {
 
-	public class OrganizationUnitItem : BaseEntity<long> {
+	public class OrganizationUnitItem : BaseEntity<int> {
 
 		public string Name { get; set; }
 		public string TypeName { get; set; }
