+++ Gravitas.Model/DomainModel/PackingTare/DAO/PackingTare.cs
@@ -1,9 +1,10 @@
 ﻿using System.ComponentModel.DataAnnotations.Schema;
+using Gravitas.Model.DomainModel.Base;
 
 namespace Gravitas.Model.DomainModel.PackingTare.DAO
 {
     [Table("PackingTare")]
-    public class PackingTare : BaseEntity<long>
+    public class PackingTare : BaseEntity<int>
     {
         public string Title { get; set; }
         public float Weight { get; set; }
