+++ Gravitas.Model/DomainModel/PreRegistration/DAO/PreRegisterCompany.cs
@@ -1,9 +1,10 @@
 using System.ComponentModel.DataAnnotations.Schema;
+using Gravitas.Model.DomainModel.Base;
 
 namespace Gravitas.Model.DomainModel.PreRegistration.DAO
 {
     [Table("PreRegisterCompany")]
-    public class PreRegisterCompany : BaseEntity<long>
+    public class PreRegisterCompany : BaseEntity<int>
     {
         public string Email { get; set; }
         public bool AllowToAdd { get; set; }
