+++ Gravitas.Model/DomainModel/Queue/DAO/QueueItemPriority.cs
@@ -1,8 +1,9 @@
 ﻿using System.Collections.Generic;
+using Gravitas.Model.DomainModel.Base;
 
 namespace Gravitas.Model
 {
-    public class QueueItemPriority : BaseEntity<long>
+    public class QueueItemPriority : BaseEntity<int>
     {
         public string Description { get; set; }
