+++ Gravitas.Model/DomainModel/Ticket/DAO/TicketStatus.cs
@@ -1,8 +1,9 @@
 using System.Collections.Generic;
+using Gravitas.Model.DomainModel.Base;
 
 namespace Gravitas.Model {
 	
-	public partial class TicketStatus : BaseEntity<long> {
+	public partial class TicketStatus : BaseEntity<int> {
 
 		public TicketStatus() {
 			TicketSet = new HashSet<Ticket>();
