+++ Gravitas.Platform.Web/ViewModel/OrganizationUnit/Detail/OrganizationUnitDetailVm.cs
@@ -1,9 +1,10 @@
 using Gravitas.Model;
+using Gravitas.Model.DomainModel.Base;
 
 // ReSharper disable once CheckNamespace
 namespace Gravitas.Platform.Web.ViewModel {
 
-	public class OrganizationUnitDetailVm : BaseEntity<long> {
+	public class OrganizationUnitDetailVm : BaseEntity<int> {
 
 		public long UnitTypeId { get; set; }
 		public string UnitTypeName { get; set; }
